Partial support to digital signature verification added
This commit updates viveksjain's branch to a newer version of PDF.js.
It currently supports validation of some common digital signatures,
such as the ones produced by Adobe Acrobat Pro, through the use of
submodules.

ASN1.js and PKI.js specific versions added as dependencies.

"WebCrypto is not available error" fixed. The first five lines of code that
use the "window" variable need to be changed to "in_window" on PKI.js.

This change enables the validation and shows a box containing results of
the process.